% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_explainer_list}
\alias{make_explainer_list}
\title{Create a List of Explainers from Models}
\usage{
make_explainer_list(model1, ..., data, y, verbose = FALSE)
}
\arguments{
\item{model1}{The primary model for which an explainer will be created.}

\item{...}{Additional models for which explainers will be created.}

\item{data}{The dataset used that will be used for explaining the models (creating profiles).}

\item{y}{Response variable in the dataset.}

\item{verbose}{If TRUE, verbose output will be printed during explanation.}
}
\value{
A list of model explainers for the provided models.
}
\description{
This function creates a list of model explainers using the \code{DALEX::explain()} function.
}
\examples{
# train models
library(DALEX)
library(ranger)
library(randomForest)
library(gbm)
model1 <- ranger(survived ~
                 gender + age + class + embarked + fare + sibsp + parch,
                 data = titanic_imputed, classification = TRUE)
model2 <- glm(survived ~
              gender + age + class + embarked + fare + sibsp + parch,
              data = titanic_imputed, family = binomial)
model3 <- randomForest(factor(survived) ~
                       gender + age + class + embarked + fare + sibsp + parch,
                       data = titanic_imputed, ntree=100)
model4 <- gbm(survived ~
              gender + age + class + embarked + fare + sibsp + parch,
              data = titanic_imputed, distribution = "bernoulli")
explainers <- make_explainer_list(model1, model2, model3, model4,
              data = titanic_imputed[, -8], y = titanic_imputed$survived)
explainers

}
