% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_summary_matrix}
\alias{plot_summary_matrix}
\title{Plot Summary Matrix}
\usage{
plot_summary_matrix(res)
}
\arguments{
\item{res}{A list containing the results obtained from the \code{rashomon_detect()} function.}
}
\value{
A matrix plot visualizing \code{Rashomon_DETECT} results.
}
\description{
This function generates a summary matrix plot that visualizes various aspects of the model comparison results. It includes bar plots for pairwise distances, dissimilarity plots, and performance plots.
}
