% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rashomon_detect.R
\name{rashomon_detect}
\alias{rashomon_detect}
\title{Detect \code{k} Most Different Models in Provided Set}
\usage{
rashomon_detect(
  explainers_list,
  performance_measure = NULL,
  sorted = FALSE,
  k = 2,
  pdi_method_numerical = "pdi",
  pdi_method_categorical = "vector_distance",
  comparison = "last",
  include_categorical_variables = TRUE,
  N = NULL,
  variable_splits = NULL
)
}
\arguments{
\item{explainers_list}{A list of models wrapped in explainers using the \code{DALEX::explain()} function (\code{make_explainer_list()} function can be helpful). All models should be classification or regression models.}

\item{performance_measure}{Performance measure used to evaluate model performance on data in explainers. For regression, options include \code{"mse"}, \code{"rmse"}, \code{"r2"}, or \code{"mad"}. For classification, options include \code{"auc"}, \code{"accuracy"}, \code{"f1"}, \code{"recall"}, or \code{"precision"}. Defaults are \code{"mse"} and \code{"auc"}.}

\item{sorted}{Whether models are sorted by performance. If not, sorting will be performed based on performance on data in explainers.}

\item{k}{Number of models to be detected as most different.}

\item{pdi_method_numerical}{Method for comparing profiles for numerical variables. Options are \code{"pdi"}, \code{"derivative l2"}, \code{"l2"}, or a custom function that takes three numerical vectors (profile 1, profile 2, and variable values) and returns a distance.}

\item{pdi_method_categorical}{Method for comparing profiles for categorical variables. Options are \code{"vector_distance"} or a custom function that takes three numerical vectors (profile 1, profile 2, and variable values) and returns a distance.}

\item{comparison}{Comparison method: \code{"all"} if the next model should be selected based on comparison with all previously selected models; \code{"last"} if it should be compared only with the last selected model.}

\item{include_categorical_variables}{Whether to include categorical variables for comparison.}

\item{N}{Number of data points used for calculating profiles, passed to the \code{DALEX::model_profile()} function.}

\item{variable_splits}{Variable splits used for calculating profiles, passed to the \code{DALEX::model_profile()} function.}
}
\value{
A list containing explainers, calculated performances, profiles, dissimilarities (distances) and indices of selected models.
}
\description{
This function performs a \code{Rashomon_DETECT} algorithm on a given set of models,
aiming to identify the \code{k} most different models based on selected parameters.
It returns a list containing performance values, model profiles, selected models,
and calculated dissimilarities.
}
\examples{
# train models
library(DALEX)
library(ranger)
library(randomForest)
library(gbm)
model1 <- ranger(survived ~
                 gender + age + class + embarked + fare + sibsp + parch,
                 data = titanic_imputed, classification = TRUE)
model2 <- glm(survived ~
              gender + age + class + embarked + fare + sibsp + parch,
              data = titanic_imputed, family = binomial)
model3 <- randomForest(factor(survived) ~
                       gender + age + class + embarked + fare + sibsp + parch,
                       data = titanic_imputed, ntree=100)
model4 <- gbm(survived ~
              gender + age + class + embarked + fare + sibsp + parch,
              data = titanic_imputed, distribution = "bernoulli")
explainers <- make_explainer_list(model1, model2, model3, model4,
              data = titanic_imputed[, -8], y = titanic_imputed$survived)
res <- rashomon_detect(explainers, k=3, comparison = "last")
res

plot_summary_matrix(res)

}
